firewall4: enable mac ranges for rule
nft supports handling mac ranges and therefore this commit changes fw4
to support that feature. The src_mac is now allowed to be a address with
CIDR notation. If no CIDR notation is configured, the old logic is
applied.

So this is now possible:
```
option src_mac '00:11:AA:BB:CC:DD/24'
```

Also negation with '!' at the beginning to block every MAC not in the
range is possible.

Signed-off-by: Christian Korber <ck@dev.tdt.de>