firewall4: enable mac ranges for rule
nft supports handling mac ranges and therefore this commit changes fw4
to support that feature. The src_mac is now allowed to be a range of two
addresses. If no range is given, the old logic is applied.

So this is now possible:
```
option src_mac '00:11:AA:00:00:00-00:11:AA:FF:FF:FF'
```
In addition to the original:
```
option src_mac '00:11:AA:00:00:00'
```

This is done by changing the regex to parse for an additional MAC address
if '-' is matched after the first MAC address.

Negative matching using exclamation mark is supported.

Signed-off-by: Christian Korber <ck@dev.tdt.de>